+++ src/schemas/project/ProjectSchema.cpp
@@ -17,149 +17,151 @@
 
 #include <utility>
 
-namespace enjin::sdk::project {
+using namespace enjin::sdk::graphql;
+using namespace enjin::sdk::models;
+using namespace enjin::sdk::project;
 
 ProjectSchema::ProjectSchema(std::unique_ptr<http::IHttpClient> http_client,
                              std::shared_ptr<utils::LoggerProvider> logger_provider)
         : shared::SharedSchema(std::move(http_client), "project", std::move(logger_provider)) {
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::advanced_send_asset(AdvancedSendAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::advanced_send_asset(AdvancedSendAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::approve_enj(ApproveEnj request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::approve_enj(ApproveEnj request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::approve_enj_max(ApproveMaxEnj request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::approve_enj_max(ApproveMaxEnj request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::AccessToken>> ProjectSchema::auth_player(AuthPlayer request) {
-    return send_request_for_one<models::AccessToken>(request);
+std::future<GraphqlResponse<AccessToken>> ProjectSchema::auth_player(AuthPlayer request) {
+    return send_request_for_one<AccessToken>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::AccessToken>> ProjectSchema::auth_project(AuthProject request) {
-    return send_request_for_one<models::AccessToken>(request);
+std::future<GraphqlResponse<AccessToken>> ProjectSchema::auth_project(AuthProject request) {
+    return send_request_for_one<AccessToken>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::bridge_asset(BridgeAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::bridge_asset(BridgeAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::bridge_assets(BridgeAssets request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::bridge_assets(BridgeAssets request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::bridge_claim_asset(BridgeClaimAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::bridge_claim_asset(BridgeClaimAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::complete_trade(CompleteTrade request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::complete_trade(CompleteTrade request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::create_asset(CreateAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::create_asset(CreateAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::AccessToken>> ProjectSchema::create_player(CreatePlayer request) {
-    return send_request_for_one<models::AccessToken>(request);
+std::future<GraphqlResponse<AccessToken>> ProjectSchema::create_player(CreatePlayer request) {
+    return send_request_for_one<AccessToken>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::create_trade(CreateTrade request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::create_trade(CreateTrade request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>>
-ProjectSchema::decrease_max_melt_fee(DecreaseMaxMeltFee request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::decrease_max_melt_fee(DecreaseMaxMeltFee request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>>
-ProjectSchema::decrease_max_transfer_fee(DecreaseMaxTransferFee request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::decrease_max_transfer_fee(DecreaseMaxTransferFee request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<bool>> ProjectSchema::delete_player(DeletePlayer request) {
+std::future<GraphqlResponse<bool>> ProjectSchema::delete_player(DeletePlayer request) {
     return send_request_for_one<bool>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Player>> ProjectSchema::get_player(GetPlayer request) {
-    return send_request_for_one<models::Player>(request);
+std::future<GraphqlResponse<Player>> ProjectSchema::get_player(GetPlayer request) {
+    return send_request_for_one<Player>(request);
 }
 
-std::future<graphql::GraphqlResponse<std::vector<models::Player>>> ProjectSchema::get_players(GetPlayers request) {
-    return send_request_for_many<models::Player>(request);
+std::future<GraphqlResponse<std::vector<Player>>> ProjectSchema::get_players(GetPlayers request) {
+    return send_request_for_many<Player>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Wallet>> ProjectSchema::get_wallet(GetWallet request) {
-    return send_request_for_one<models::Wallet>(request);
+std::future<GraphqlResponse<Wallet>> ProjectSchema::get_wallet(GetWallet request) {
+    return send_request_for_one<Wallet>(request);
 }
 
-std::future<graphql::GraphqlResponse<std::vector<models::Wallet>>> ProjectSchema::get_wallets(GetWallets request) {
-    return send_request_for_many<models::Wallet>(request);
+std::future<GraphqlResponse<std::vector<Wallet>>> ProjectSchema::get_wallets(GetWallets request) {
+    return send_request_for_many<Wallet>(request);
 }
 
-std::future<graphql::GraphqlResponse<bool>> ProjectSchema::invalidate_asset_metadata(InvalidateAssetMetadata request) {
+std::future<GraphqlResponse<bool>> ProjectSchema::invalidate_asset_metadata(InvalidateAssetMetadata request) {
     return send_request_for_one<bool>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::melt_asset(MeltAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::melt_asset(MeltAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::message(Message request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::message(Message request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::mint_asset(MintAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::mint_asset(MintAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::release_reserve(ReleaseReserve request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::release_reserve(ReleaseReserve request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::reset_enj_approval(ResetEnjApproval request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::reset_enj_approval(ResetEnjApproval request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::send_asset(SendAsset request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::send_asset(SendAsset request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::send_enj(SendEnj request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::send_enj(SendEnj request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_approval_for_all(SetApprovalForAll request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_approval_for_all(SetApprovalForAll request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_melt_fee(SetMeltFee request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_melt_fee(SetMeltFee request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_transfer_fee(SetTransferFee request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_transfer_fee(SetTransferFee request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_transferable(SetTransferable request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_transferable(SetTransferable request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_uri(SetUri request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_uri(SetUri request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<models::Transaction>> ProjectSchema::set_whitelisted(SetWhitelisted request) {
-    return send_request_for_one<models::Transaction>(request);
+std::future<GraphqlResponse<Transaction>> ProjectSchema::set_whitelisted(SetWhitelisted request) {
+    return send_request_for_one<Transaction>(request);
 }
 
-std::future<graphql::GraphqlResponse<bool>> ProjectSchema::unlink_wallet(UnlinkWallet request) {
+std::future<GraphqlResponse<bool>> ProjectSchema::unlink_wallet(UnlinkWallet request) {
     return send_request_for_one<bool>(request);
 }
 
+std::future<GraphqlResponse<Transaction>> ProjectSchema::update_name(UpdateName request) {
+    return send_request_for_one<Transaction>(request);
 }